+++ Polymorphism/Hierarchy.Tests/AnimalTests.cs
@@ -1,0 +1,13 @@
+using Microsoft.VisualStudio.TestTools.UnitTesting;
+
+namespace Hierarchy.Tests
+{
+    [TestClass]
+    public class AnimalTests
+    {
+        [TestMethod]
+        public void TestMethod1()
+        {
+        }
+    }
+}
