+++ Polymorphism/Hierarchy.Tests/CatTests.cs
@@ -1,0 +1,13 @@
+using Microsoft.VisualStudio.TestTools.UnitTesting;
+
+namespace Hierarchy.Tests
+{
+    [TestClass]
+    public class CatTests
+    {
+        [TestMethod]
+        public void TestMethod1()
+        {
+        }
+    }
+}
