+++ Polymorphism/Hierarchy.Tests/MammalTests.cs
@@ -1,0 +1,13 @@
+using Microsoft.VisualStudio.TestTools.UnitTesting;
+
+namespace Hierarchy.Tests
+{
+    [TestClass]
+    public class MammalTests
+    {
+        [TestMethod]
+        public void TestMethod1()
+        {
+        }
+    }
+}
