+++ Polymorphism/Hierarchy/Animals/Mammal.cs
@@ -1,4 +1,4 @@
-﻿namespace Hierarchy
+﻿namespace Hierarchy.Animals
 {
     public abstract class Mammal : Animal
     {
