+++ Polymorphism/Hierarchy/Hierarchy.csproj
@@ -5,4 +5,8 @@
         <TargetFramework>netcoreapp5.0</TargetFramework>
     </PropertyGroup>
 
+    <ItemGroup>
+      <Folder Include="Exceptions\" />
+    </ItemGroup>
+
 </Project>
