+++ Polymorphism/Polymorphism.sln
@@ -24,7 +24,9 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ScooterRental.Tests", "Test
 		{192D97CD-3CE9-4580-9779-3B1261E9DFBF} = {192D97CD-3CE9-4580-9779-3B1261E9DFBF}
 	EndProjectSection
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "VendingMachine.Tests", "VendingMachine.Tests\VendingMachine.Tests.csproj", "{CC0DA5B6-37BF-479E-BC48-4E71F3263C23}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "VendingMachine.Tests", "VendingMachine.Tests\VendingMachine.Tests.csproj", "{CC0DA5B6-37BF-479E-BC48-4E71F3263C23}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Hierarchy.Tests", "Hierarchy.Tests\Hierarchy.Tests.csproj", "{20EC374E-13EE-401B-9916-465EBD6459BF}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -72,6 +74,10 @@ Global
 		{CC0DA5B6-37BF-479E-BC48-4E71F3263C23}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{CC0DA5B6-37BF-479E-BC48-4E71F3263C23}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{CC0DA5B6-37BF-479E-BC48-4E71F3263C23}.Release|Any CPU.Build.0 = Release|Any CPU
+		{20EC374E-13EE-401B-9916-465EBD6459BF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{20EC374E-13EE-401B-9916-465EBD6459BF}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{20EC374E-13EE-401B-9916-465EBD6459BF}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{20EC374E-13EE-401B-9916-465EBD6459BF}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
